refactor(cli): extract unity support utilities
